requirements.txt
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.11.0
bce-python-sdk==0.9.46
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorlog==6.10.1
filelock==3.20.0
fsspec==2025.9.0
future==1.0.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
imagesize==1.4.1
modelscope==1.31.0
networkx==3.5
numpy==2.3.4
opencv-contrib-python==4.10.0.84
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.3.0
paddlepaddle==3.2.0
paddlex==3.3.3
pandas==2.3.3
pillow==12.0.0
prettytable==3.16.0
protobuf==6.33.0
psutil==7.1.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.12.2
pydantic-core==2.41.4
pypdfium2==4.30.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
requests==2.32.5
ruamel-yaml==0.18.15
ruamel-yaml-clib==0.2.14
safetensors==0.6.2
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
wcwidth==0.2.14

# FastAPI and Web Framework
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Environment and Settings
pydantic-settings==2.7.1

# Additional utilities
python-dotenv==1.0.1
